requirements.txt
azure-common==1.1.28
azure-core==1.24.2
azure-search-documents==11.2.2
azure-storage-blob==12.12.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.0
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
